Exclude thinking from intermediate query responses
Without doing this, RAG can't work with models that include their
reasoning in the output, because these responses are being re-used
as reformulated prompts. The whole thing gets confused when there
is reasoning in there. I don't think it's necessarily the right
approach, but it gets thinks working for me with a self-hosted
Deepseek model.